% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMetrics_Calc.R
\name{AllMetrics_Calc}
\alias{AllMetrics_Calc}
\title{Calculate all the associated drug response metrices}
\usage{
AllMetrics_Calc(All_GDSC_GR_data, metric, Metrics_headers, writeFile = T,
  fileName)
}
\arguments{
\item{All_GDSC_GR_data}{A data frame with CellLine, Drug, Concentration, GR/NDR/PI and Relative Viability of all GDSC data}
}
\value{
A data frame with all the GR/NDR/PI associated metrics
}
\description{
Fit sigmoid curves and compute all associated GR metrics (GR50,GEC50,GRmax,GRinf,GRhill,GRAOC,GRR2) and other metrics NDR / PI for GDSC
Significance of curve fits assessed based on the conditions specified in Marc Hafner's paper
Concentrations are capped to the highest concentration value, when the GEC50s are 10 times more than the highest concentration; Same also applies, when the p-values resulting from the significance of fits is >0.5
}
