% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Associations_overlap_percentages.R
\name{Associations_overlap_percentages}
\alias{Associations_overlap_percentages}
\title{Association overlap between different metrics}
\usage{
Associations_overlap_percentages(metric1, metric2, metric1_name, metric2_name)
}
\arguments{
\item{metric1}{Conventional metric}

\item{metric2}{GR/NDR metric}
}
\value{
A dataframe with the Association thresholds and Percentage overlap
}
\description{
Calculation of overlaps between the conventional and GR/NDR metrics based on varying thresholds
}
