% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_GDSC.R
\name{calculate_relativeViability}
\alias{calculate_relativeViability}
\title{Calculate Relative viabilities for GDSC data}
\usage{
calculate_relativeViability(RawData)
}
\arguments{
\item{RawData}{read \code{\link{AllDrugs_CellLines_GDSC}} for data format.}

\item{Drug_Conc}{data.frame with Drug, CellLine, and Concetration columns.}

\item{writeFile}{(default FALSE) write results into file}

\item{outFile}{destination of results}
}
\value{
vector with relative viabilities
}
\description{
Calculate relative viabilities (Relative Viability = (Raw intensity - Positive Control) / (Negative Control - Positive Control))
}
