% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellLines_filter.R
\name{CellLines_filter}
\alias{CellLines_filter}
\title{Filter the cell lines based on the criteria described in Hafner's publication}
\usage{
CellLines_filter(Gemcitabine_GR_Ref, Gemcitabine_GR_var, Gemcitabine_IC_Ref,
  Gemcitabine_IC_Var, GDSC_R2, GCSI_published_Gemcitabine, writeFile = F,
  outFile = "Calculated_Values/GDSC_final_CellLines_toCalculate_GR.txt")
}
\arguments{
\item{Gemcitabine_GR_Ref}{A dataframe with cell lines and Reference values computed from Gemcitabine responses (GCSI data)}

\item{Gemcitabine_GR_var}{A dataframe with cell lines and variance values computed from Gemcitabine responses (GCSI data)}

\item{Gemcitabine_IC_Ref}{A dataframe with cell lines and Reference values computed from Gemcitabine responses (GDSC data)}

\item{Gemcitabine_IC_Var}{A dataframe with cell lines and variance values computed from Gemcitabine responses (GDSC data)}
}
\value{
A vector with final set of cell lines to be used for GR calculations
}
\description{
Filter the cell lines based on the criteria described in Hafner's publication
}
