% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances_DrugTargets.R
\name{Distances_DrugTargets}
\alias{Distances_DrugTargets}
\title{Calculation of Distances to Drug Targets}
\usage{
Distances_DrugTargets(metric, write.File = T, FileName)
}
\arguments{
\item{metric}{ANOVA results corresponding to a specific metric}
}
\value{
A dataframe with the drug, FDR gene (Gene that is identified as significant in FDR and has FDR<10%), Original target and distances
}
\description{
Drug targets retrieved from GDSC and CHEMBL
}
