% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugResponsePredictions_Ml_models.R
\name{DR_Models}
\alias{DR_Models}
\title{Generate drug response prediction models and use them to predict external test sets}
\usage{
DR_Models(LigandDesc_LNIC50, MutCNV)
}
\arguments{
\item{MutCNV}{A dataframe with principal component scores of the mutation / CNV status of the cell lines}

\item{LigandDesc}{A dataframe with ECFP4 descriptors for drugs and LNIC50 responses corresponding to different cell lines}
}
\value{
A dataframe with the various test set performances
}
\description{
Function to generate drug response prediction models and predict external test sets
}
