% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCSI_Conventional_GRMetrics_published.R
\docType{data}
\name{GCSI_Conventional_GRMetrics_published}
\alias{GCSI_Conventional_GRMetrics_published}
\title{Published GR values of Genentech screening data}
\format{A data frame with 6167 rows and 4 variables:
\describe{
  \item{Drug}{Name of the drug}
  \item{CellLine}{Name of the CellLine}
  \item{GCSI_AUC}{Area Under the Curve}
  \item{GCSI_GR_AOC}{Area Over the Curve}
  }}
\source{
\url{http://www.grcalculator.org/grbrowser/}
}
\usage{
GCSI_Conventional_GRMetrics_published
}
\description{
A dataset containing the GR and conventional metrics of GCSI screening data (Data downloaded from \link {http://www.grcalculator.org/grbrowser/})
}
\keyword{datasets}
