% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCSI_GR.R
\name{GCSI_GR}
\alias{GCSI_GR}
\title{Calculate GR values for Gemcitabine (Reference compound);}
\usage{
GCSI_GR(Gemcitabine_GCSI_Via_Doses, CellDivisionRates, Assay_Time,
  writeFile = T,
  outFile = "Calculated_Values/Gemcitabine_GCSI_GR_allCellLines.txt")
}
\arguments{
\item{Gemcitabine_GCSI_Via_Doses}{Dose response data of Gemcitabine}

\item{CellDivisionRates}{Division times of the cell lines}

\item{writeFile}{(default TRUE) Save results to a file}

\item{outFile}{file destination for results}
}
\value{
data.frame with CellLine, Conc  and GR columns
}
\description{
Division times and viabilities from GCSI used
}
