% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSC_DrugIDs_Names_mapped.R
\docType{data}
\name{GDSC_DrugIDs_Names_mapped}
\alias{GDSC_DrugIDs_Names_mapped}
\title{GDSC Drug ids and Drug names mapping}
\format{A data frame with 267 rows and 2 variables:
\describe{
  \item{DRUG_ID}{IDs of the drug}
  \item{DRUG_NAME}{Name of the drug}
  }}
\source{
\url{https://www.cancerrxgene.org/downloads}
}
\usage{
GDSC_DrugIDs_Names_mapped
}
\description{
Annotations of the drugs downloaded from GDSC \url{https://www.cancerrxgene.org/downloads}
}
\details{
A dataframe with drug ids and the drug names
}
\keyword{datasets}
