% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSC_GR.R
\name{GDSC_GR}
\alias{GDSC_GR}
\title{Calculate GR for GDSC}
\usage{
GDSC_GR(GDSC_finalCellLines, IGDSC_ref, GR_GCSI_ref, GDSC_via_complete,
  writeFile = T, outFile = "Calculated_Values/AllGR_GDSCdata.txt")
}
\arguments{
\item{GDSC_finalCellLines}{Set of cell lines which have passed the quality checks and for which GR metrics is to be calculated}

\item{IGDSC_ref}{Reference values of cell lines calculated from relative viabilities of Gemcitabine (GDSC data)}

\item{GR_GCSI_ref}{Reference values of cell lines calculated from GRs of Gemcitabine (GCSI data)}

\item{GDSC_via_complete}{Relative viabilities of GDSC cell lines}
}
\value{
A data frame with cellLine, Drug, Concentration and GR values for final set of GDSC cell lines
}
\description{
Calculate GR for GDSC
}
