% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSC_KnownAssociations.R
\docType{data}
\name{GDSC_KnownAssociations}
\alias{GDSC_KnownAssociations}
\title{Well-known gene-drug associations}
\format{A data frame with 20 rows and 1 variables:
\describe{
  \item{Ident_Drug_Feature}{Drugs and the Mutation/CNV status of genes}
  }}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/27397505}
}
\usage{
GDSC_KnownAssociations
}
\description{
Established biomarker information downloaded from Iorio et al., 2016 \url{https://www.ncbi.nlm.nih.gov/pubmed/27397505}
}
\details{
A dataframe with drugs and the corresponding gene feature
}
\keyword{datasets}
