% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genentech_GCSI_DivisionTimes_CellLines.R
\docType{data}
\name{Genentech_GCSI_DivisionTimes_CellLines}
\alias{Genentech_GCSI_DivisionTimes_CellLines}
\title{Division times of cell lines from Genentech Screening data}
\format{A data frame with 408 rows and 2 variables:
\describe{
  \item{CellLine}{Name of the cell line}
  \item{Doubling_Time}{Cell doubling time}
  }}
\source{
\url{https://www.nature.com/articles/nature17987#supplementary-information}
}
\usage{
Genentech_GCSI_DivisionTimes_CellLines
}
\description{
A dataset containing the doubling time of celllines
}
\keyword{datasets}
