% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterquartileMean.R
\name{InterquartileMean}
\alias{InterquartileMean}
\title{Calculate interquartile means for Gemcitabine data across different concentrations}
\usage{
InterquartileMean(Gemcitabine_GR_filtered, type, metric = c("Rel_via", "GR"),
  writeFile = T, outFile = paste("Calculated_Values/Gemcitabine_", type,
  "_allCellLines.txt", sep = ""))
}
\arguments{
\item{Gemcitabine_GR_filtered}{Filtered Gemcitabine response data between concentration 0.1 and 32Mm}

\item{type}{GCSI_GRRef or GDSC_ICRef (Specify whether the interquartile means are to be calculated for the Gemcitabine data from GCSI or GDSC)}

\item{metric}{GR or Rel_via (Specify whether GR or Relative Viability values are to be used for interquartile men calculations)}
}
\value{
A dataframe with cell lines and interquartile means of GRs or Relative viabilities
}
\description{
Interquartile means correspond to the reference values used for calculating GRs for other cell lines and drugs in GDSC
}
